+++ files/Server/Unix/p4/common/bin/weekly_sync_replica.sh
@@ -42,7 +42,10 @@ fi
 check_vars
 set_vars
 
+# override setting of checkpoint.log from set_vars
 LOGFILE=$LOGS/sync_replica.log
+
+check_uid
 
 $P4BIN -u $P4USER -p ${SSL_PREFIX}${P4MASTER}:${P4MASTERPORTNUM} login < /p4/common/bin/adminpass > /dev/null 2>&1
 JOURNALNUM=`$P4BIN -u $P4USER -p ${SSL_PREFIX}${P4MASTER}:${P4MASTERPORTNUM} counter journal`
@@ -56,11 +59,13 @@ log "Starting weekly_sync_replica.sh"
 
 # You must set up a public keypair using "ssh-keygen -t rsa" in order for this to work.
 # You need to paste your CLIENT ~/.ssh/id_rsa.pub contents into the REMOTE ~/ssh/authorized_keys file. 
-rsync -avz --delete ${OSUSER}@${P4MASTER}:$CHECKPOINTS/ $CHECKPOINTS > $LOGFILE 2>&1
-rsync_exit_code=$?
+if [[ "$SHAREDDATA" == "FALSE" ]]; then
+   rsync -avz --delete ${OSUSER}@${P4MASTER}:$CHECKPOINTS/ $CHECKPOINTS > $LOGFILE 2>&1
+   rsync_exit_code=$?
 
-if [[ $rsync_exit_code -ne 0 ]]; then
-   die "Error: Failed to pull /p4/$SDP_INSTANCE/checkpoints from host $P4MASTER.  The rsync exit code was: $rsync_exit_code.  Aborting."
+   if [[ $rsync_exit_code -ne 0 ]]; then
+      die "Error: Failed to pull $CHECKPOINTS from host $P4MASTER.  The rsync exit code was: $rsync_exit_code.  Aborting."
+   fi
 fi
 
 recreate_weekly_offline_db_files
@@ -98,7 +103,10 @@ replay_journals_to_offline_db
 
 check_disk_space
 remove_old_logs
+export CHECKPOINTS=${P4HOME}/checkpoints.rep
+remove_old_checkpoints_and_journals
+
 $P4BIN -p ${SSL_PREFIX}${P4MASTERPORTNUM} pull -lj >> $LOGFILE
 
-log "End $P4SERVER Checkpoint"
+log "End $P4SERVER weekly sync replica log."
 mail_log_file "$HOSTNAME $P4SERVER Weekly sync replica log."
